Add click off functionality on rendered menus:
Add click off functionality for both subscription and filter menus that take both menus away when a click occurs off the menus themselves. This was done by implementing react-onClickOutside library, as well as a conditional statement for the Filter menus to track what is being clicked on.

Change (e) to (event) on functions in SearchFilter for deeper clarity.

Fix small bug in App.jsx where initial loading caused No Search Results Found to render caused by shouldComponentMount. Adding a condition to check for text in the searchbar was the solution.