Add automated weekly email fucntionality.
Create createEmailTemplate file which holds functionality to create the HTML email template to be sent to mailchimp. The function gets called from listings.js every week when the database gets cleaned, with an array containing the company name, title, and url of all new companies added to the database.

Create sendEmailTemplate file that holds the sendTemplate post requests that takes the HTML email template with new job listings and creates the new email at mailchimp. This post request calls another createCampaign which creates the campaign at mailchimp with the email template, and finally that request calls sendCampaign that sends the campaign to the weekly job listings list.

Clean up some code and add promise in listings.js in updateDatabase function that waits for all of the async database functions to finish running before combineCompanyListings calls getEmailList in it's promise chain.